### 2024-07-15 #248 @kyleniemeyer
Hi @nicoguaro, I'm also on the JOSE editorial board (and a JOSS associate editor in chief) and would prefer not to review in addition to those duties.
