### 2024-08-01 #248 @nicoguaro
@chennachaos, would you like to review this submission for JOSE? It concerns an asynchronous course that "is developed to help learners with a basic understanding of fluid dynamics and CFD bridge the knowledge gap toward the effective usage of CFD on modern [...] resources."
