### 2024-08-01 #252 @editorialbot
[REVIEW]: ARC4CFD: Learning how to leverage High-Performance Computing with Computational Fluid Dynamics
**Submitting author:** @j6hickey (<a href="http://orcid.org/0000-0002-6944-3964">Jean-Pierre Hickey</a>)
**Repository:** https://github.com/ARC4CFD/arc4cfd.github.io
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
