### 2024-08-01 #252 @nicoguaro
@j6hickey, could you please add a license to the repository so we can proceed with the review?

> **License info**:
>
> 🔴 Failed to discover a valid open source license
