### 2024-08-23 #252 @chennachaos
Hi @nicoguaro, Regarding authorship, the submitting author, @j6hickey, made only 3 out of 431 commits. Is this contribution sufficient? Are there any guidelines on the amount of contribution by the submitting author?
