### 2024-09-03 #252 @dortiz5
Dear @j6hickey, thank you for this fantastic course and the related repository.
The course has different learning sections with examples, nice images, explanatory videos, quizzes, and a glossary. Also, it was very interesting, pedagogical, and self-contained. More precisely, I found section 2 very beneficial to reduce the gap between small-scale and large-scale simulations, which is the main scope of the course.
I have some suggestions regarding the checklist:
1- Including a version indicator of the course would be helpful if further relevant changes are made.
2- Although the course is self-contained, it would be beneficial for students and other instructors to include a brief description in the documentation of the repository of the programs and packages in Windows, Mac, and Linux to be used during the course. Also, include the recommended minimal version to ensure the best experience and avoid further problems.
3- Finally, include brief instructions in the repository's documentation on making third-party contributions, such as including the link (https://arc4cfd.github.io/contact/) to send you the comments.
Best regards,
David
