### 2024-10-04 #252 @j6hickey
We would like to thank @dortiz5 for the wonderful feedback that was provided. We have addressed all the suggestions that helped to strengthen the course. Here are some details on what was done:
1- We included a version tracker into the git repository. This will allow the course authors to flag any major changes to the course content. At the moment, as the course is on its first release, we have set the version counter to V1.0.
2- This was a very good point. We designed the course for users of the Digital Research Alliance of Canada clusters which share a common software stack. As the course was also intended for non-Canadian users and users without access to HPC, we have added a discussion in the Overview page about these points. We also clarified the versions of the open-source toolkit used in the course (namely Gmsh, OpenFoam, and SU2) as well as the minimal version requirement for the installation of the codes. We have also referred the learners back to this page if they want to install these tools locally.
3 – We have added a short description in the Overview about how the community can contribute to this on-going course development. It is intended that a synchronous version of this course will be offered yearly in June as part of the Compute Ontario Summer School, therefore any changes and improvements will likely occurs around this time of year.
