### 2024-10-04 d872a51
Update jose_pdf.yml to try to fix deprecated workflow
