### 2024-11-15 #252 @j6hickey
@nicoguaro @chennachaos I was wondering if there were any updates on the proposed revisions needed for this manuscript. Thanks for all your help!

Best,
Jean-Pierre
