### 2024-11-23 #252 @chennachaos
Hi @j6hickey, This is a well-developed course on HPC for CFD! I will certainly recommend it to those interested.

I have the following comments.

1. Add a link to the course page (https://arc4cfd.github.io/) in the repository's description.
2. The statement "The course was specifically developed for use on the Digital Research Alliance of Canada HPC clusters but access to an HPC cluster is not essential for completing the course. The students can install the software and run the test cases on their local computer.", is misleading as the students cannot follow the HPC-related tasks, e.g., the tasks in Section 1 are not possible to run a laptop. So, modify this appropriately, clarifying what the users need to do if they want to run the examples on the local machines using multiple processors.
3. In the paper, add the software and example problems covered in the course. Also, add appropriate references for them.
