+++ paper.md
@@ -35,7 +35,7 @@ Computational Fluid Dynamics (CFD) is a field of computational physics that has 
 
 
 # Statement of need
-Although most science and engineering programs offer CFD courses, which have become standard part of the curriculum in mechanical, civil, chemical, and aerospace engineering, these courses generally focus on the fundamental understanding of the physics, modelling, and numerics as well as the practical usage of CFD tools. At the end of a typical undergraduate CFD course, the student has learned to use a CFD tool to solve small-scale problems, understand the modelling assumptions, quantify numerical errors, and visualize the results. The CFD problems used in the tutorials of these courses are designed to run on students' personal or local workstations. The jump from small-scale CFD usage on a local workstation to the effective utilization of this same CFD tool, on a much larger problem size, on modern HPC systems is nontrivial and requires additional specific training. Although generic HPC and parallel computing training is widely available, targeted training material for CFD users on HPC systems is not readily found. Thus, these learners must either: a) "translate" the knowledge from generic HPC training material to the field of CFD, or b) rely on mentorship and external help to effectively utilize these computational resources.  As these CFD tools are continually improving, they are increasingly being adopted to complement experimental campaigns, inform the design of experiments, or help solve theoretical fluid dynamics problems. This underserved user base is seeking focused training materials for the effective utilization of CFD on HPC systems, as they do not have the same mentorship opportunities or access to expertise as graduate students. ARC4CFD was developed to provide a CFD-specific training material for the effective use of HPC systems in advanced research computing architectures. The course is built in an asynchronous format for a broader adoption, and many hands-on problems are integrated to help the learner further their understanding and learn independently. We value the ability to freely share and exchange with the community, and thus the focus is placed on an entirely open-source toolset, from meshing, solving, and visualizing the results. Although this course is intended for use on remote HPC systems, and more specifically on the Digital Research Alliance of Canada clusters, a user can easily follow on a local multiprocessor computer, or any other HPC system.
[…]
