+++ paper.bib
@@ -22,3 +22,44 @@ url = {https://doi.org/10.21105/jose.00021},
   year={2002},
   publisher={The Royal Society}
 }
+
+@article{Geuzaine:2009,
+  title={Gmsh: a three-dimensional finite element mesh generator with built-in pre-and post-processing facilities},
+  author={Geuzaine, Christophe and Remacle, Jean-Fran{\c{c}}ois and others},
+  journal={International journal for numerical methods in engineering},
+  volume={79},
+  number={11},
+  pages={1309--1331},
+  year={2009}
+}
+
+@article{Weller:1998,
+  title={A tensorial approach to computational continuum mechanics using object-oriented techniques},
+  author={Weller, Henry G and Tabor, Gavin and Jasak, Hrvoje and Fureby, Christer},
+  journal={Computers in physics},
+  volume={12},
+  number={6},
+  pages={620--631},
+  year={1998},
+  publisher={American Institute of Physics}
+}
+
+@incollection{palacios:2013,
+	title = {Stanford {University} {Unstructured} ({SU}{\textasciicircum}2): {An} open-source integrated computational environment for multi-physics simulation and design},
+	url = {https://doi.org/10.2514/6.2013-287},
+	booktitle = {51st {AIAA} {Aerospace} {Sciences} {Meeting} including the {New} {Horizons} {Forum} and {Aerospace} {Exposition}},
+	publisher = {American Institute of Aeronautics and Astronautics},
+	author = {Palacios, Francisco and Alonso, Juan and Duraisamy, Karthikeyan and Colonno, Michael and Hicken, Jason and Aranake, Aniket and Campos, Alejandro and Copeland, Sean and Economon, Thomas and Lonkar, Amrita and Lukaczyk, Trent and Taylor, Thomas},
+	year = {2013},
+	doi = {10.2514/6.2013-287}
+}
+
+@article{ahrens:2005,
+  title={Paraview: An end-user tool for large-data visualization},
+  author={Ahrens, James and Geveci, Berk and Law, Charles and Hansen, C and Johnson, C},
+  journal={The visualization handbook},
+  volume={717},
+  pages={50038--1},
+  year={2005},
+  publisher={Citeseer}
+}
