+++ paper.md
@@ -57,7 +57,7 @@ The course is expected to take about 16 hours to complete and is divided into th
 - 2.6 Running the CFD on HPC.
 - 2.7 A posteriori analysis. 
 
-The structure of Section 2 provides the learner with a comprehensive overview of a standard workflow for a CFD problem on HPC systems. A summative example, which is based on the simulation of a backward facing step, is integrated into the classes of Section 2 and follows the main steps of the CFD workflow.
+The structure of Section 2 provides the learner with a comprehensive overview of a standard workflow for a CFD problem on HPC systems. A summative example, which is based on the simulation of a backward facing step, is integrated into the classes of Section 2 and follows the main steps of the CFD workflow. This summative example is used to highlight the methodology to estimate the computational cost, setup the simulation, mesh, run, and postprocess. Additional examples include boundary layer (both laminar and turbulent) that used to highlight the near-wall resolution requirements, a cylinder in crossflow to illustrate a sample mesh generation process.
 
 There are five learning outcomes for ARC4CFD. At the end of the course, the learner should be able to:
 
@@ -70,14 +70,14 @@ There are five learning outcomes for ARC4CFD. At the end of the course, the lear
 The learning outcomes for each class are presented and reinforced with questions at the end of each lecture. 
 
 # Philosophy of learning
-ARC4CFD is an asynchronous online course that accompanies the learner in developing skills to effectively utilize Advanced Research Computing for Computational Fluid Dynamics. The course is built using an *integrative learning* approach in which concepts from a wide range of scientific disciplines are brought together (e.g., parallel computing, programming, numerics, fluid dynamics, CFD) and directly put into practice through hands-on examples and quizzes. This course is built around the usage of an entirely open-source meshing, solving, and visualization toolset. This is a key feature of the course that we hope will encourage a broad adoption of the training material. For this reason, the course is licensed under the Creative Commons BY-NC-SA license, and all the original files (e.g., svg files or Python scripts) are included in the repository.  The main open-source toolkits used include: Gmsh (meshing), SU2 (CFD solver), OpenFoam (CFD solver), Paraview (visualization), and Python (postprocessing and analysis). Gmsh is a widely used meshing that can be used to generate structured and unstructured meshes for both solid and fluid mechanics [@Geuzaine:2009]. The two different open-source CFD frameworks are used for better reach among the various communities; both solvers rely on a classical finite volume formulation. OpenFoam [@Weller:1998]  is a toolbook to solve problems in continuum mechanics and is one of the most used CFD solvers, due versatile in developing complex numerical solvers. SU2, on the other hand, has a stronger aerospace heritage with a growing user base across various disciplines [@palacios:2013]. Finally, visualization of the results relies on Paraview [@ahrens:2005].
+ARC4CFD is an asynchronous online course that accompanies the learner in developing skills to effectively utilize Advanced Research Computing for Computational Fluid Dynamics. The course is built using an *integrative learning* approach in which concepts from a wide range of scientific disciplines are brought together (e.g., parallel computing, programming, numerics, fluid dynamics, CFD) and directly put into practice through hands-on examples and quizzes. This course is built around the usage of an entirely open-source meshing, solving, and visualization toolset. This is a key feature of the course that we hope will encourage a broad adoption of the training material. For this reason, the course is licensed under the Creative Commons BY-NC-SA license, and all the original files (e.g., svg files or Python scripts) are included in the repository.  
 
[…]
