### 2024-12-06 #252 @j6hickey
Thanks @chennachaos for the great comments to help improve the course and paper. Here are the rebuttals. We have addressed all the points that were raised:

**Add a link to the course page ( https://arc4cfd.github.io/) in the repository's description.**

Good point. We have added the course webpage at a number of locations including the main README files in the repository.

**The statement "The course was specifically developed for use on the Digital Research Alliance of Canada HPC clusters but access to an HPC cluster is not essential for completing the course. The students can install the software and run the test cases on their local computer.", is misleading as the students cannot follow the HPC-related tasks, e.g., the tasks in Section 1 are not possible to run a laptop. So, modify this appropriately, clarifying what the users need to do if they want to run the examples on the local machines using multiple processors.**

We agree with this statement and we have modified the course accordingly.   As we are focusing on developing HPC skills, we have de-emphasized the ability to conduct local tests. As section 2 is CFD focused, we nonetheless wanted to allow students that don’t have access to the HPC resources to use our training modules. We modified the Overview (and other locations) with the new description that reads:

“This course uses open-source CFD tools on modern high-performance computing systems. The course was  specifically developed for use on the Digital Research Alliance of Canada HPC clusters, although users of other similar high-performance clusters (e.g. linux-based systems, using SLURM workload manager and module system) can use the course with minimal changes. Nonetheless, access to an HPC cluster is not essential for completing some parts of the course. The students can install the open-source software and run the test cases on their local computer; the only requirement is to use a multi-core processor. There will be slight differences in installation procedures, specific command lines and visualization. Users on other HPC clusters are invited to reach out to their local help desk for assistance.”

**In the paper, add the software and example problems covered in the course. Also, add appropriate references for them.**

	We have added a description of the software used along with the appropriate references:
“ The main open-source toolkits used include: Gmsh (meshing), SU2 (CFD solver), OpenFoam (CFD solver), Paraview (visualization), and Python (postprocessing and analysis). Gmsh is a widely used meshing that can be used to generate structured and unstructured meshes for both solid and fluid mechanics [@Geuzaine:2009]. The two different open-source CFD frameworks are used for better reach among the various communities; both solvers rely on a classical finite volume formulation. OpenFoam [@Weller:1998]  is a toolbook to solve problems in continuum mechanics and is one of the most used CFD solvers, due versatile in developing complex numerical solvers. SU2, on the other hand, has a stronger aerospace heritage with a growing user base across various disciplines [@palacios:2013]. Finally, visualization of the results relies on Paraview [@ahrens:2005].

Additionally, the description of the example problems has been added:
“A summative example, which is based on the simulation of a backward facing step, is integrated into the classes of Section 2 and follows the main steps of the CFD workflow. This summative example is used to highlight the methodology to estimate the computational cost, setup the simulation, mesh, run, and postprocess. Additional examples include boundary layer (both laminar and turbulent) that used to highlight the near-wall resolution requirements, a cylinder in crossflow to illustrate a sample mesh generation process.”
