### 2024-12-07 #252 @chennachaos
HI @nicoguaro I am satisfied with the changes. I recommend the library for publication.
