### 2024-12-07 #252 @nicoguaro
@j6hickey, can you add the missing DOIs?

Also, I think that the paper is missing a verb in line 50.
