### 2024-12-07 bace128
added doi and fixed line 50 in paper
