+++ paper.bib
@@ -18,6 +18,7 @@ url = {https://doi.org/10.21105/jose.00021},
   journal={Philosophical Transactions of the Royal Society of London. Series A: Mathematical, Physical and Engineering Sciences},
   volume={360},
   number={1795},
+  doi = {10.1098/rsta.2002.0990},
   pages={1211--1225},
   year={2002},
   publisher={The Royal Society}
@@ -30,6 +31,7 @@ url = {https://doi.org/10.21105/jose.00021},
   volume={79},
   number={11},
   pages={1309--1331},
+  doi ={10.1002/nme.2579},
   year={2009}
 }
 
@@ -39,6 +41,7 @@ url = {https://doi.org/10.21105/jose.00021},
   journal={Computers in physics},
   volume={12},
   number={6},
+  doi={10.1063/1.168744},
   pages={620--631},
   year={1998},
   publisher={American Institute of Physics}
@@ -61,5 +64,6 @@ url = {https://doi.org/10.21105/jose.00021},
   volume={717},
   pages={50038--1},
   year={2005},
+  doi = {10.1016/b978-012387582-2/50038-1},
   publisher={Citeseer}
 }
