### 2024-12-07 #252 @j6hickey
@nicoguaro Thanks for your assistance. I have added the DOIs and cleaned up the sentence in line 50. Thanks!
