### 2024-12-07 #252 @nicoguaro
@j6hickey, I think that your work would benefit of using two licenses, one for the code (it might be MIT) and the one you are already using CC4-NC-SA. What do you think?
