### 2024-12-08 #252 @j6hickey
@nicoguaro Thanks. Good idea. We added the MIT licence to cover the software side of the work.
