### 2024-12-09 #252 @nicoguaro
@j6hickey can you create a new version of the repo and archive it at Zenodo?
