### 2024-12-09 #252 @j6hickey
Yes, I have uploaded to Zenodo. To finalize the archival process, I would need the DOI and other information on the paper. Here is the DOI of the Zenodo repository: https://doi.org/10.5281/zenodo.14340786
