### 2024-12-09 #252 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
