### 2024-12-09 #252 @nicoguaro
@labarba, this paper is ready to move forward.
