### 2024-12-24 #252 @labarba
I found the paper to be a little verbose in places, and when running into a couple of actual typos decided to go back and do a little copy editing. See my PR here:
https://github.com/ARC4CFD/arc4cfd.github.io/pull/2
