### 2024-12-24 #252 @labarba
License mismatch: the Zenodo repository shows a Creative Commons Attribution Non Commercial 4.0 International, but elsewhere it is shown as CC BY NC SA.

While I mention that, may I raise a question about your choice of a non-commercial license? There are a few reasons why this could be problematic. See: https://gw-ospo.github.io/oss-licensing/license_your_own_work.html#empowering-others-to-build-upon-your-work
