### 2025-01-03 #252 @j6hickey
Thanks for the comments on the repository. We were unaware of the potential issues arising from the non-commercial licence. We have decided to use a CC BY 4.0 licence instead. These changes have now been applied to the git repo and a new version of the repository was uploaded to the zenodo archive:   https://doi.org/10.5281/zenodo.14593527. These changes should address the licence mismatch concern.
