### 2025-04-16 #252 @labarba
Hi, thanks for the ping. Pulling up the repository, I see only one [tagged release](https://github.com/ARC4CFD/arc4cfd.github.io/tags). After the review and revision process, the authors should issue a new tagged release, and the version is updated here.
