### 2025-04-16 #252 @labarba
copy editing PR not merged
https://github.com/ARC4CFD/arc4cfd.github.io/pull/2
