### 2025-05-04 #252 @labarba
The Zenodo archive is v2 — did you also want to issue a new tagged release in the GitHub repository and update the version here?
