### 2025-05-05 #252 @labarba
The Zenodo archive still has a mismatched license: it shows CC-BY-NC and MIT, while the materials are CC-BY...
