### 2025-05-05 #252 @j6hickey
@labarba Sorry about that. Made the changes to Zenodo archive. It is now CC BY 4.0 and should be consistent with the Git repo.
