### 2025-05-05 #252 @labarba
Uhm... I still see the NC license (tried in a private window to confirm it's not a browser cache issue)
